+++ addOns/alertFilters/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Added support for parameter regex, attack and evidence strings and regexes (Issue 5574)
 
 ## [8] - 2019-06-07